refactor(web): extract use-tool-state.ts from SpatialEditor
The tool/mode state (tool, toolChosenByUserRef, initialToolAppliedRef,
the initial-tool-once effect, contextMenu, longPressPulse,
openContextMenuAtRef) moves verbatim into its own hook, called where
its first declaration stood — ahead of every other cluster still
pending extraction.

editor-state-surface.test.ts's sources glob gains use-tool-state.ts
per its own registration contract. Mutation-checked: temporarily
dropping the entry turned the ledger's "state no longer held" check
red on tool/contextMenu/longPressPulse, confirming the guard reaches
this hook rather than passing vacuously.
