+++ apps/web/src/components/spatial-editor/editor-state-surface.test.ts
@@ -40,6 +40,7 @@ const sources = import.meta.glob(
     './SpatialEditor.tsx',
     './use-editor-measurements.ts',
     './use-file-seam-scene.ts',
+    './use-tool-state.ts',
     './use-viewport-controls.ts',
   ],
   { query: '?raw', import: 'default', eager: true },
